Fix pytest imports
