+++ trustpoint/request/tests/test_authentication_extended.py
@@ -18,13 +18,13 @@ from devices.models import (
     OnboardingPkiProtocol,
 )
 from pki.util.keys import KeyGenerator
-from request.authentication.base import (
+from request.authentication import (
     ClientCertificateAuthentication,
-    CmpSharedSecretAuthentication,
     IDevIDAuthentication,
-    UsernamePasswordAuthentication,
 )
-from request.request_context import RequestContext
+from request.authentication.cmp import CmpSharedSecretAuthentication
+from request.authentication.est import UsernamePasswordAuthentication
+from request.request_context import BaseRequestContext
 
 
 @pytest.mark.django_db
@@ -48,7 +48,7 @@ class TestUsernamePasswordAuthenticationExtended:
         device.save()
         
         auth = UsernamePasswordAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.est_username = device.common_name
         context.est_password = 'anypassword'
         
@@ -71,7 +71,7 @@ class TestUsernamePasswordAuthenticationExtended:
         )
         
         auth = UsernamePasswordAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.est_username = device.common_name
         context.est_password = 'anypassword'
         
@@ -81,7 +81,7 @@ class TestUsernamePasswordAuthenticationExtended:
     def test_authenticate_exception_during_lookup(self) -> None:
         """Test authentication handles exceptions during device lookup."""
         auth = UsernamePasswordAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.est_username = 'test-user'
         context.est_password = 'test-pass'
         
@@ -100,7 +100,7 @@ class TestClientCertificateAuthenticationExtended:
     def test_authenticate_no_client_certificate(self) -> None:
         """Test authentication returns None when no client certificate provided."""
         auth = ClientCertificateAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.client_certificate = None
         
         result = auth.authenticate(context)
@@ -138,7 +138,7 @@ class TestClientCertificateAuthenticationExtended:
         ).sign(private_key.as_crypto(), hashes.SHA256())
         
         auth = ClientCertificateAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.client_certificate = cert
         
         with pytest.raises(ValueError, match="Client certificate not recognized"):
@@ -161,7 +161,7 @@ class TestClientCertificateAuthenticationExtended:
         cert = issued_cred.credential.get_certificate()
         
         auth = ClientCertificateAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.client_certificate = cert
         
         # Mock the is_valid_domain_credential to return False
@@ -179,7 +179,7 @@ class TestCmpSharedSecretAuthentication:
     def test_authenticate_non_cmp_protocol(self) -> None:
         """Test authentication raises error when protocol is not CMP."""
         auth = CmpSharedSecretAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.protocol = 'est'  # Wrong protocol
         
         with pytest.raises(ValueError, match="CMP shared secret authentication requires CMP protocol"):
@@ -188,7 +188,7 @@ class TestCmpSharedSecretAuthentication:
     def test_authenticate_no_parsed_message(self) -> None:
         """Test authentication raises error when no parsed message."""
         auth = CmpSharedSecretAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.protocol = 'cmp'
         context.parsed_message = None
         
@@ -198,7 +198,7 @@ class TestCmpSharedSecretAuthentication:
     def test_authenticate_invalid_message_type(self) -> None:
         """Test authentication raises error with invalid message type."""
         auth = CmpSharedSecretAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.protocol = 'cmp'
         context.parsed_message = Mock()  # Not a PKIMessage
         context.parsed_message.__class__.__name__ = 'SomeOtherType'
@@ -214,7 +214,7 @@ class TestIDevIDAuthentication:
     def test_authenticate_no_raw_message(self) -> None:
         """Test authentication returns None when no raw_message."""
         auth = IDevIDAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.raw_message = None
         
         result = auth.authenticate(context)
@@ -230,7 +230,7 @@ class TestIDevIDAuthentication:
         domain = device_instance['domain']
         
         auth = IDevIDAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.raw_message = Mock()  # Non-None raw message
         context.domain = domain
         
@@ -250,7 +250,7 @@ class TestIDevIDAuthentication:
         domain = device_instance['domain']
         
         auth = IDevIDAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.raw_message = Mock()  # Non-None raw message
         context.domain = domain
         
@@ -270,7 +270,7 @@ class TestIDevIDAuthentication:
         domain = device_instance['domain']
         
         auth = IDevIDAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.raw_message = Mock()  # Non-None raw message
         context.domain = domain
         
@@ -292,7 +292,7 @@ class TestIDevIDAuthentication:
         device.save()
         
         auth = IDevIDAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.raw_message = Mock()  # Non-None raw message
         context.domain = None
         
@@ -326,7 +326,7 @@ class TestCompositeAuthentication:
         
         # Try username/password authentication
         auth = UsernamePasswordAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.est_username = device.common_name
         context.est_password = 'test-password-123'
         
@@ -352,7 +352,7 @@ class TestCompositeAuthentication:
                 onboarding_config.save()
                 
                 auth = UsernamePasswordAuthentication()
-                context = Mock(spec=RequestContext)
+                context = Mock(spec=BaseRequestContext)
                 context.est_username = device.common_name
                 context.est_password = 'onboarding-password-123'
                 
@@ -371,7 +371,7 @@ class TestCompositeAuthentication:
             device.save()
             
             auth = UsernamePasswordAuthentication()
-            context = Mock(spec=RequestContext)
+            context = Mock(spec=BaseRequestContext)
             context.est_username = device.common_name
             context.est_password = 'onboarding-password-123'
             
@@ -408,7 +408,7 @@ class TestAuthenticationEdgeCases:
         device.save()
         
         auth = UsernamePasswordAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.est_username = 'device-special-@#$'
         context.est_password = 'p@ssw0rd!#$%^&*()'
         
@@ -433,7 +433,7 @@ class TestAuthenticationEdgeCases:
         device.save()
         
         auth = UsernamePasswordAuthentication()
-        context = Mock(spec=RequestContext)
+        context = Mock(spec=BaseRequestContext)
         context.est_username = device.common_name.upper()  # Wrong case
         context.est_password = 'test-password'
