+++ trustpoint/request/tests/test_pki_message_parser.py
@@ -7,17 +7,11 @@ from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.primitives.asymmetric import rsa
 from pki.models import DomainModel
 
-from request.message_parser.base import (
-    CertProfileParsing,
-    CmpMessageParser,
-    CmpPkiMessageParsing,
-    CompositeParsing,
-    DomainParsing,
-    EstCsrSignatureVerification,
-    EstMessageParser,
-    EstPkiMessageParsing,
-)
-from request.request_context import RequestContext
+from request.message_parser import CmpMessageParser, EstMessageParser
+from request.message_parser.base import CertProfileParsing, CompositeParsing, DomainParsing
+from request.message_parser.cmp import CmpPkiMessageParsing
+from request.message_parser.est import EstAuthorizationHeaderParsing, EstCsrSignatureVerification, EstPkiMessageParsing
+from request.request_context import BaseRequestContext
 
 
 class TestEstPkiMessageParsing:
@@ -25,7 +19,7 @@ class TestEstPkiMessageParsing:
 
     def test_parse_pem_csr_success(self, test_csr_fixture):
         """Test parsing a valid PEM-encoded CSR."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = test_csr_fixture.get_pem()
         mock_context.raw_message = mock_raw_message
@@ -39,7 +33,7 @@ class TestEstPkiMessageParsing:
 
     def test_parse_base64_der_with_newlines_success(self, test_csr_fixture):
         """Test parsing a valid Base64-encoded DER CSR with newlines."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = test_csr_fixture.get_base64_der_with_newlines()
         mock_context.raw_message = mock_raw_message
@@ -53,7 +47,7 @@ class TestEstPkiMessageParsing:
 
     def test_parse_base64_der_csr_success(self, test_csr_fixture):
         """Test parsing a valid Base64-encoded DER CSR."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = test_csr_fixture.get_base64_der()
         mock_context.raw_message = mock_raw_message
@@ -67,7 +61,7 @@ class TestEstPkiMessageParsing:
 
     def test_parse_der_csr_success(self, test_csr_fixture):
         """Test parsing a valid raw DER CSR."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = test_csr_fixture.get_der()
         mock_context.raw_message = mock_raw_message
@@ -81,7 +75,7 @@ class TestEstPkiMessageParsing:
 
     def test_parse_missing_raw_message(self):
         """Test parsing with missing raw message."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.raw_message = None
 
         parser = EstPkiMessageParsing()
@@ -94,7 +88,7 @@ class TestEstPkiMessageParsing:
 
     def test_parse_missing_message_body(self):
         """Test parsing with missing message body."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = None
         mock_context.raw_message = mock_raw_message
@@ -109,7 +103,7 @@ class TestEstPkiMessageParsing:
 
     def test_parse_unsupported_format(self):
         """Test parsing with unsupported CSR format."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = b'This is not valid base64 data!'
         mock_context.raw_message = mock_raw_message
@@ -124,7 +118,7 @@ class TestEstPkiMessageParsing:
 
     def test_parse_exception_handling(self):
         """Test handling of parsing exceptions."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = b'-----BEGIN CERTIFICATE REQUEST-----\ntest\n-----END CERTIFICATE REQUEST-----'
         mock_context.raw_message = mock_raw_message
@@ -145,7 +139,7 @@ class TestEstCsrSignatureVerification:
 
     def test_verify_rsa_signature_success(self, test_csr_fixture):
         """Test successful RSA signature verification."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.cert_requested = test_csr_fixture.get_cryptography_object()
 
         verifier = EstCsrSignatureVerification()
@@ -154,7 +148,7 @@ class TestEstCsrSignatureVerification:
 
     def test_verify_missing_csr(self):
         """Test verification with missing CSR."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.cert_requested = None
 
         verifier = EstCsrSignatureVerification()
@@ -167,7 +161,7 @@ class TestEstCsrSignatureVerification:
 
     def test_verify_missing_hash_algorithm(self):
         """Test verification with missing hash algorithm."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_csr = Mock(spec=x509.CertificateSigningRequest)
         mock_csr.signature_hash_algorithm = None
         mock_context.cert_requested = mock_csr
@@ -182,7 +176,7 @@ class TestEstCsrSignatureVerification:
 
     def test_verify_unsupported_key_type(self):
         """Test verification with unsupported key type."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_csr = Mock(spec=x509.CertificateSigningRequest)
         mock_unsupported_key = Mock()
         mock_hash_algorithm = Mock(spec=hashes.SHA256)
@@ -201,7 +195,7 @@ class TestEstCsrSignatureVerification:
 
     def test_verify_signature_failure(self):
         """Test handling of signature verification failure."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_csr = Mock(spec=x509.CertificateSigningRequest)
         mock_rsa_key = Mock(spec=rsa.RSAPublicKey)
         mock_hash_algorithm = Mock(spec=hashes.SHA256)
@@ -224,12 +218,109 @@ class TestEstCsrSignatureVerification:
                 assert 'Failed to verify the CSR signature.' in str(e)
 
 
+class TestEstAuthorizationHeaderParsing:
+    """Test EstAuthorizationHeaderParsing class."""
+
+    def test_validate_success(self):
+        """Test successful validation and credential extraction."""
+        validator = EstAuthorizationHeaderParsing()
+        context = BaseRequestContext()
+        context.raw_message = MagicMock()
+
+        # Create valid Basic auth header
+        credentials = 'username:password'
+        encoded_credentials = base64.b64encode(credentials.encode()).decode()
+        context.raw_message.headers = {'Authorization': f'Basic {encoded_credentials}'}
+
+        validator.validate(context)
+        assert context.est_username == 'username'
+        assert context.est_password == 'password'
+
+    def test_validate_missing_authorization(self):
+        """Test that missing Authorization header passes."""
+        validator = EstAuthorizationHeaderParsing()
+        context = BaseRequestContext()
+        context.raw_message = MagicMock()
+        context.raw_message.headers = {'Other-Header': 'value'}
+
+        # Should not raise any exception
+        validator.validate(context)
+
+    def test_validate_non_basic_auth(self):
+        """Test that non-Basic auth does not pass."""
+        validator = EstAuthorizationHeaderParsing()
+        context = BaseRequestContext()
+        context.raw_message = MagicMock()
+        context.raw_message.headers = {'Authorization': 'Bearer token123'}
+
+        try:
+            validator.validate(context)
+            assert False, 'Expected ValueError to be raised'
+        except ValueError as e:
+            assert "Authorization header must start with 'Basic'." in str(e)
+
+    def test_validate_malformed_basic_auth(self):
+        """Test ValueError for malformed Basic auth credentials."""
+        validator = EstAuthorizationHeaderParsing()
+        context = BaseRequestContext()
+        context.raw_message = MagicMock()
+        context.raw_message.headers = {'Authorization': 'Basic invalid_base64!!!'}
+
+        try:
+            validator.validate(context)
+            assert False, 'Expected ValueError to be raised'
+        except ValueError as e:
+            assert "Malformed 'Authorization' header credentials" in str(e)
+
+    def test_validate_basic_auth_no_colon(self):
+        """Test ValueError for Basic auth without colon separator."""
+        validator = EstAuthorizationHeaderParsing()
+        context = BaseRequestContext()
+        context.raw_message = MagicMock()
+
+        # Valid base64 but no colon separator
+        credentials = 'usernamepassword'
+        encoded_credentials = base64.b64encode(credentials.encode()).decode()
+        context.raw_message.headers = {'Authorization': f'Basic {encoded_credentials}'}
+
+        try:
+            validator.validate(context)
+            assert False, 'Expected ValueError to be raised'
+        except ValueError as e:
+            assert "Malformed 'Authorization' header credentials" in str(e)
+
+    def test_validate_missing_raw_message(self):
+        """Test ValueError when raw_message is None."""
+        validator = EstAuthorizationHeaderParsing()
+        context = BaseRequestContext()
+        context.raw_message = None
+
+        try:
+            validator.validate(context)
+            assert False, 'Expected ValueError to be raised'
+        except ValueError as e:
+            assert 'Raw message is missing from the context' in str(e)
+
+    def test_validate_missing_headers(self):
+        """Test ValueError when headers are missing."""
+        validator = EstAuthorizationHeaderParsing()
+        context = BaseRequestContext()
+        context.raw_message = MagicMock()
+        context.raw_message.headers = None
+
+        try:
+            validator.validate(context)
+            assert False, 'Expected ValueError to be raised'
+        except ValueError as e:
+            assert 'Raw message is missing headers' in str(e)
+
+
 class TestDomainParsing:
     """Test cases for DomainParsing component."""
 
     def test_parse_domain_success(self):
         """Test successful domain parsing."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.domain_str = 'test.domain.com'
         mock_domain = Mock(spec=DomainModel)
 
@@ -246,7 +337,7 @@ class TestDomainParsing:
         It is not mandatory to have a domain at this stage (e.g. general endpoint without path segment).
         Domain can be resolved from device in authentication step.
         """
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.domain_str = None
 
         parser = DomainParsing()
@@ -254,7 +345,7 @@ class TestDomainParsing:
 
     def test_parse_domain_validation_error(self):
         """Test domain validation error handling."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.domain_str = 'invalid.domain.com'
 
         parser = DomainParsing()
@@ -265,7 +356,7 @@ class TestDomainParsing:
 
     def test_parse_domain_not_found(self):
         """Test domain not found error handling."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.domain_str = 'missing.domain.com'
 
         parser = DomainParsing()
@@ -311,7 +402,7 @@ class TestCertProfileParsing:
 
     def test_parse_cert_profile_str_success(self):
         """Test successful certificate profile string parsing."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.cert_profile_str = 'test_template'
 
         parser = CertProfileParsing()
@@ -321,7 +412,7 @@ class TestCertProfileParsing:
 
     def test_parse_missing_cert_profile_str(self):
         """Test parsing with missing certificate template."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.cert_profile_str = None
 
         parser = CertProfileParsing()
@@ -338,7 +429,7 @@ class TestCmpPkiMessageParsing:
 
     def test_parse_cmp_message_success(self):
         """Test successful CMP message parsing."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = b'cmp_message_data'
         mock_context.raw_message = mock_raw_message
@@ -354,7 +445,7 @@ class TestCmpPkiMessageParsing:
 
     def test_parse_missing_raw_message(self):
         """Test parsing with missing raw message."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_context.raw_message = None
 
         parser = CmpPkiMessageParsing()
@@ -367,7 +458,7 @@ class TestCmpPkiMessageParsing:
 
     def test_parse_missing_message_body(self):
         """Test parsing with missing message body."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = None
         mock_context.raw_message = mock_raw_message
@@ -382,7 +473,7 @@ class TestCmpPkiMessageParsing:
 
     def test_parse_decode_error(self):
         """Test handling of decode errors."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = b'invalid_cmp_data'
         mock_context.raw_message = mock_raw_message
@@ -398,7 +489,7 @@ class TestCmpPkiMessageParsing:
 
     def test_parse_type_error(self):
         """Test handling of type errors during decode."""
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
         mock_raw_message = Mock()
         mock_raw_message.body = b'invalid_cmp_data'
         mock_context.raw_message = mock_raw_message
@@ -472,7 +563,7 @@ class TestCmpMessageParser:
     def test_parse_delegation(self):
         """Test that parse method delegates to components."""
         parser = CmpMessageParser()
-        mock_context = Mock(spec=RequestContext)
+        mock_context = Mock(spec=BaseRequestContext)
 
         # Set up the mock context with required attributes for CMP parsing
         mock_context.raw_message = Mock()
